build(folders): :truck: reorganize data access folders
reorganize data access folders

remove unneeded comments and imports
