+++ src/dataAccess/getAllSecrets.js
@@ -1,5 +1,3 @@
-import Secrets from './model.js';
-
 /*
 getAllSecrets function will display
 all secrets in the Secrets table
@@ -15,15 +13,9 @@ async function getAllSecrets(table) {
             return { key: secretObj.key, value: secretObj.value };
         });
         return secretsFormatted;
-        // [
-        //   { key: 'exampleKey1', value: 'exampleValue1' },
-        //   { key: 'exampleKey2', value: 'exampleValue2' },
-        // ]
     } catch (error) {
         console.error('Error getting all secrets: ', error);
     }
 }
-
-// getAllSecrets();
 
 export default getAllSecrets;
